support histogram intervals of type double

+++ src/main/java/org/elasticsearch/search/facet/histogram/HistogramFacet.java
@@ -145,7 +145,7 @@ public interface HistogramFacet extends Facet, Iterable<HistogramFacet.Entry> {
         /**
          * The key value of the histogram.
          */
-        long getKey();
+        double getKey();
 
         /**
          * The number of hits that fall within that key "range" or "interval".
